move solver components into their own files
