+++ lib/solve/solver.rb
@@ -1,129 +1,10 @@
 module Solve
-
   # @author Jamie Winsor <jamie@vialstudios.com>
   class Solver
-    # @author Andrew Garson <andrew.garson@gmail.com>
-    # @author Jamie Winsor <jamie@vialstudios.com>
-    class VariableTable
-      attr_reader :rows
-
-      def initialize
-        @rows = Array.new
-      end
-
-      def add(package, source)
-        row = rows.detect { |row| row.package == package }
-        if row.nil?
-          @rows << Variable.new(package, source)
-        else
-          row.add_source(source)
-        end
-      end
-
-      def first_unbound
-        @rows.detect { |row| row.bound? == false }
-      end
-
-      def find_package(package)
-        @rows.detect { |row| row.package == package }
-      end
-
-      def remove_all_with_only_this_source!(source)
-        with_only_this_source, others = @rows.partition { |row| row.sources == [source] }
-        @rows = others
-        with_only_this_source
-      end
-
-      def all_from_source(source)
-        @rows.select { |row| row.sources.include?(source) }
-      end
-
-      def before(package)
-        package_index = @rows.index { |row| row.package == package }
-        (package_index == 0) ? nil : @rows[package_index - 1]
-      end
-
-      def all_after(package)
-        package_index = @rows.index { |row| row.package == package }
-        @rows[(package_index+1)..-1]
-      end
-    end
-
-    # @author Andrew Garson <andrew.garson@gmail.com>
-    # @author Jamie Winsor <jamie@vialstudios.com>
-    class Variable
-      attr_reader :package
-      attr_reader :value
-      attr_reader :sources
-
-      def initialize(package, source)
-        @package = package
-        @value = nil
-        @sources = Array(source)
-      end
-
-      def add_source(source)
-        @sources << source
-      end
-
-      def last_source
-        @sources.last
-      end
-
-      def bind(value)
-        @value = value
-      end
-
-      def unbind
-        @value = nil
-      end
-
-      def bound?
-        !@value.nil?
-      end
-
-      def remove_source(source)
-        @sources.delete(source)
-      end
-    end
-
-    # @author Andrew Garson <andrew.garson@gmail.com>
-    # @author Jamie Winsor <jamie@vialstudios.com>
-    class ConstraintTable
-      attr_reader :rows
-
-      def initialize
-        @rows = Array.new
-      end
-
-      def add(package, constraint, source)
-        @rows << ConstraintRow.new(package, constraint, source)
-      end
-
-      def constraints_on_package(package)
-        @rows.select do |row|
-          row.package == package
-        end.map { |row| row.constraint }
-      end
-
-      def remove_constraints_from_source!(source)
-        @rows.reject! { |row| row.source == source }
-      end
-    end
-
-    # @author Andrew Garson <andrew.garson@gmail.com>
-    # @author Jamie Winsor <jamie@vialstudios.com>
-    class ConstraintRow
-      attr_reader :package
-      attr_reader :constraint
-      attr_reader :source
-
-      def initialize(package, constraint, source)
-        @package = package
-        @constraint = constraint
-        @source = source
-      end
-    end
+    autoload :VariableTable, 'solve/solver/variable_table'
+    autoload :Variable, 'solve/solver/variable'
+    autoload :ConstraintTable, 'solve/solver/constraint_table'
+    autoload :ConstraintRow, 'solve/solver/constraint_row'
 
     class << self
       # Create a key to identify a demand on a Solver.
@@ -347,7 +228,6 @@ module Solve
       end
 
       def backtrack(unbound_variable)
-
         previous_variable = variable_table.before(unbound_variable.package)
 
         if previous_variable.nil?
